Fix NSS paths for Unix. This one fixes multilib Fedora.
This has been a long standing issue, see:

- https://github.com/ctt-gob-es/clienteafirma/issues/27
- https://github.com/ctt-gob-es/clienteafirma/issues/16

We now include /usr/lib/firefox- as a valid path by default.

Rationale:

- The code never worked in the obvious way: it never seeked valid
  Firefox installations [1]

- In particular, behavior for including /usr/lib/firefox-xx.x.x was
  never there, so integrators will already have their workarounds and
  this is not needed

- The workaround at #27 may still be used somewhere

[1] As pointed out at
https://github.com/ctt-gob-es/clienteafirma/issues/27#issuecomment-488402089
the searchLastFirefoxVersion function will always return an empty
string, so the only additional directory searched for was
"/usr/lib/firefox-".

If it's needed in the future it can be taken from this commit's parent
revision.